chore(api domain): refactor spec processing
This PR separates the internal fields in the API Spec out into discrete types, moving the internal implementation details farther from the types that exist to unmarshall user input.

It looked weird and wrong (but does work) to have both the user-input Request and internal request types in the same package, so I moved the user types into schemas.