fix: fixes around nist 800 rev 5 controls (#347)
## Description

Fixes around shared UUID that came up when we migrated from NIST 800 Rev
4 to NIST 800 Rev 5. This PR changes the way that the mappings are held
in ephemeral memory to allow the Lula UI to display mappings that share
UUID and `control_id`

Notable Changes:
- The ephemeral state (for mappings), mainly `map[string]interface{}`
uses a new key, hash, which is the sha256sum of the mapping.
- The checksums must be created on the backend. There is webcrypto but
it is not supported in all browsers
- There was a general agreement not to create any breaking changes, ie,
the mappings can remain as it and when being read from the filesystem
the hash will be generated.

## Related Issue

Fixes #340 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://github.com/defenseunicorns/lula/blob/main/CONTRIBUTING.md)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>